tweet_consumer/requirements.txt
kafka-python==2.0.2
influxdb_client==1.14.0
vadersentiment==3.3.2
